task-creator/requirement.txt
uuid
yt-dlp
fake-useragent
functions-framework
google-api-python-client
pymongo
google-cloud-pubsub
google-cloud-tasks
google-cloud-storage 
google-cloud-speech 
google-cloud-translate 
pymongo 
moviepy 
pydub